move: tab items constants
